requirements.txt
imgaug==0.4.0
imutils==0.5.4
matplotlib==3.7.1
ncnn==1.0.20240410
numpy==1.24.4
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
paddlepaddle==2.6.1
pandas==2.2.2
Pillow==10.4.0
plotly==5.14.1
protobuf==3.20.0
python-dotenv==1.0.1
torch==2.0.1
torch_summary==1.4.5
tqdm==4.65.2
ultralytics==8.2.38
x2paddle==1.4.1
onnxruntime==1.18.0
streamlit==1.36.0